CRUD on Webhooks, exceptions handler

+++ src/main/java/com/computools/service/ConfigurationsServiceImpl.java
@@ -1,14 +1,20 @@
 package com.computools.service;
 
+import com.amazonaws.services.dynamodbv2.datamodeling.QueryResultPage;
+import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.computools.repository.ConfigurationsRepository;
+import com.computools.repository.WebhooksRepository;
 import com.computools.repository.table.Configurations;
 import com.computools.service.dto.ConfigurationsDto;
+import com.computools.service.dto.QueryResultPageConfigurationsDto;
 import com.computools.service.mapper.ConfigurationsMapper;
+import com.computools.service.mapper.QueryResultPageConfigurationsMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -20,7 +26,13 @@ public class ConfigurationsServiceImpl implements ConfigurationsService {
     private ConfigurationsRepository configurationsRepository;
 
     @Autowired
+    private WebhooksRepository webhooksRepository;
+
+    @Autowired
     private ConfigurationsMapper configurationsMapper;
+
+    @Autowired
+    private QueryResultPageConfigurationsMapper queryResultPageConfigurationsMapper;
 
     @Override
     public ConfigurationsDto save(ConfigurationsDto dto) {
@@ -44,14 +56,17 @@ public class ConfigurationsServiceImpl implements ConfigurationsService {
     }
 
     @Override
-    public List<ConfigurationsDto> findByHashKey(Long tenantNo) {
-        List<ConfigurationsDto> result = configurationsRepository
-                .findByHashKey(tenantNo).stream().map(c->configurationsMapper.toDto(c)).collect(Collectors.toList());
+    public QueryResultPageConfigurationsDto findByHashKey(Long tenantNo, Map<String, AttributeValue> lastEvaluatedKey) {
+        QueryResultPage<Configurations> page =
+                configurationsRepository.findByHashKey(tenantNo, lastEvaluatedKey);
+        QueryResultPageConfigurationsDto result = queryResultPageConfigurationsMapper.toDto(page);
         return result;
     }
 
     @Override
     public void deleteByCompositeKey(Long tenantNo, String configurationId) {
+        if (webhooksRepository.existsByCompositeIndex(tenantNo, configurationId))
+            throw new IllegalStateException("Configurations with associated webhooks cannot be deleted! Delete webhooks first!");
         configurationsRepository.deleteByCompositeKey(tenantNo, configurationId);
     }
 }
